refactor: extract BanditsAtGate event into bandits_event.js and wire import in TownRuntime
